Move reused sprinkles @sargassum-world/stimulated, remove TypeScript
