+++ web/app/package.json
@@ -4,21 +4,17 @@
   "private": true,
   "scripts": {
     "build": "rollup -c",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "format": "prettier --write ./src/**/*.{js,svelte,html,ts}",
-    "lint": "tsc --noEmit && prettier --write ./src/**/*.{js,svelte,html,ts} && eslint src/**/*.{js,ts,svelte}",
-    "lint:fix": "tsc --noEmit && prettier --write ./src/**/*.{js,svelte,html,ts} && eslint --fix src/**/*.{js,ts,svelte}"
+    "check": "svelte-check",
+    "format": "prettier --write ./src/**/*.{js,svelte,html}",
+    "lint": "prettier --write ./src/**/*.{js,svelte,html} && eslint src/**/*.{js,svelte}",
+    "lint:fix": "prettier --write ./src/**/*.{js,svelte,html} && eslint --fix src/**/*.{js,svelte}"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
-    "@rollup/plugin-typescript": "^8.0.0",
     "@sargassum-world/styles": "^0.2.0",
-    "@tsconfig/svelte": "^2.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
     "bulma": "^0.9.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.10.0",
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-svelte3": "^3.2.1",
     "prettier": "^2.4.1",
@@ -34,16 +30,13 @@
     "svelte": "^3.0.0",
     "svelte-check": "^2.0.0",
     "svelte-dark-mode": "^2.0.0",
-    "svelte-preprocess": "^4.0.0",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.0"
+    "svelte-preprocess": "^4.0.0"
   },
   "dependencies": {
     "@fontsource/atkinson-hyperlegible": "^4.5.1",
     "@fontsource/oxygen-mono": "^4.5.0",
     "@hotwired/turbo": "^7.0.1",
-    "async-mutex": "^0.3.2",
-    "stimulus": "2.0.0",
-    "stimulus-use": "^0.41.0"
+    "@sargassum-world/stimulated": "^0.2.9",
+    "stimulus": "2.0.0"
   }
 }
