+++ web/app/src/main-eager.js
@@ -1,0 +1,5 @@
+import { ThemeToggle } from '@sargassum-world/stimulated/util';
+
+ThemeToggle.onLoad();
+
+export {};
