--- web/app/src/main-eager.ts
@@ -1,5 +1,0 @@
-import * as ThemeToggle from './sprinkles/theme-toggle';
-
-ThemeToggle.onLoad();
-
-export {};
