--- web/app/src/sprinkles/turbo-cache.controller.js
@@ -1,11 +1,0 @@
-import * as Turbo from '@hotwired/turbo';
-import { Controller } from 'stimulus';
-
-export default class extends Controller {
-  clear(e) {
-    // This assumes the controller is attached to a form!
-    e.preventDefault();
-    Turbo.clearCache();
-    e.target.submit();
-  }
-}
